+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.1] - 27/08/2025
+
+### Fixed
+- Bet display when player is playing insurance.
+
+### Changed
+- Reduced dealer blackjack checking delay.
+
 ## [1.4.0] - 23/08/2025
 
 ### Added